+++ TennisClub.DAL/Entities/MemberRole.cs
@@ -6,7 +6,7 @@ namespace TennisClub.DAL.Entities
     {
         public int Id { get; set; }
         public int MemberId { get; set; }
-        public byte RoleId { get; set; }
+        public int RoleId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? EndDate { get; set; }
         public Member MemberNavigation { get; set; }
